baseline/automix/pyproject.toml
[project]
name = "automix"
version = "0.1.0"
description = "Automix Router - Self-verification based LLM routing"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "sentencepiece>=0.1.99",
    "numpy>=1.21",
    "pandas>=1.5",
    "scikit-learn>=1.2",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "pydantic>=2.0",
    "litellm>=1.0",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/automix"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
